compositor|refactor: Give StaleSnapshotError its own module and a message
Moves `StaleSnapshotError` out of `foldSourceTiers` into `config/StaleSnapshotError.ts` and has it take a message, so a second refusal can raise it: a config whose folded source list has moved away from the one a snapshot was captured over. The two refusals mean one thing to a caller -- locate and capture again -- and a caller handling one wants to handle both.

The unlocated package and its base directory stay on the error as optional fields, set only where an unlocated package is what went stale.
