+++ packages/compositor/src/index.ts
@@ -9,7 +9,7 @@ export { computeClosure } from './closure/computeClosure.ts';
 export type { ArtifactRead, EdgeContribution, EdgeContributor } from './closure/EdgeContributor.ts';
 export type { EdgeGraph } from './closure/EdgeGraph.ts';
 export type { SourceResolution } from './config/foldSourceTiers.ts';
-export { foldSourceTiers, StaleSnapshotError } from './config/foldSourceTiers.ts';
+export { foldSourceTiers } from './config/foldSourceTiers.ts';
 export type { TierFile } from './config/loadConfig.ts';
 export { loadConfig, TierFileSchema } from './config/loadConfig.ts';
 export type { LocatePackageOptions } from './config/locatePackage.ts';
@@ -17,6 +17,8 @@ export { locatePackage } from './config/locatePackage.ts';
 export type { LocateSourcePackagesOptions, PackageLocation } from './config/locateSourcePackages.ts';
 export { locateSourcePackages } from './config/locateSourcePackages.ts';
 export { resolveSources } from './config/resolveSources.ts';
+export type { UnlocatedPackage } from './config/StaleSnapshotError.ts';
+export { StaleSnapshotError } from './config/StaleSnapshotError.ts';
 export { ConsistencyError } from './consistency/ConsistencyError.ts';
 export {
   ARTIFACT_ID_PLACEHOLDER,
